Move function with metadata
